+++ src/control_plane/shared_infrastructure/postgres/mod.rs
@@ -1,11 +1,15 @@
+mod plan_postgres_project_resources;
 mod postgres_logical_resource_plan;
 mod postgres_plan_error;
+mod postgres_project_resources;
 mod postgres_shared_instance_plan;
 mod postgres_shared_instance_plan_options;
 mod provision_postgres_logical_resource;
 
+pub(crate) use plan_postgres_project_resources::plan_postgres_project_resources;
 pub(crate) use postgres_logical_resource_plan::PostgresLogicalResourcePlan;
 pub(crate) use postgres_plan_error::PostgresPlanError;
+pub(crate) use postgres_project_resources::PostgresProjectResources;
 pub(crate) use postgres_shared_instance_plan::PostgresSharedInstancePlan;
 pub(crate) use postgres_shared_instance_plan_options::PostgresSharedInstancePlanOptions;
 pub(crate) use provision_postgres_logical_resource::provision_postgres_logical_resource;